+++ lib/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: UTF-8 -*-
+
+"""
+    This package contains ...
+"""
+
+from lib.localization import *
+from lib.kalman_localization import *
